A R script for generating synthetic data sets and analysing them with the Lasso model.
First, an instance generater is set-up.
Second, we define a latin-hypercube (LH) scheme for efficient sampling of parameter space.
Third, for each parameter combination of the LH sampling scheme, one instance (data set) is generated.
Fifth, for each instance we tune the parameter Lambda of the Lasso. 
Based on the optimum parameter of Lambda, we record generalization error, sparsity, and total computation time (including tuning) for the selected model.
This script depends on the following secondary R script files:
properties.R: here the paths to the working directory (workingDir) and derivated directories (dataDir, plotDir, codeDir)
have to be defined.